app/requirements.txt
tornado
requests
nose
bson
pymongo
pillo